Initial commit of new benchmark library.
The current version of the benchmark library introduces more overhead than is
desirable each test iteration. This is the initial commit of a new version that
attempts to fix this.

This version introduces a new test API without the State object. It also removes
a bunch of functionality including:

- Support for different regex libraries.
- Removes test reporter customization.
- Removes color output.
- and more!

The plan is to re-add this functionality and then merge master and this branch.