RFC: Replace reporters with JSON and tooling.
The only supported reporter in the library outputs JSON. Tooling has
been added to generate CSV and Tabular reports from the JSON output.

This change is incomplete and is an RFC. Known issues: lack of testing
for tooling, some documentation needs to be updated, no colour support
in tabular output tooling.

Enables #441, #190.