Rename BUILD to BUILD.bazel to work around case-insentivity.
The rationale for making this change is docmented here:
    https://github.com/google/benchmark/pull/501

Briefly speaking, some test bots on TravisCI do not have case sensitive
filesystems.  So the build will fail to create the `build directory`.

It does not look like Travis is going to resolve this problem on their side
anytime soon, yet.  So I am making this change to pass the test.