Rename StrCat to StringCat
StrCat is a macro in Shlwapi.h that conflicts with StrCat in string_util.h:

   #defines StrCat lstrcatA

When Shlwapi.h is included in sysinfo.cc and all references to StrCat are
renamed to lstrcatA in string_util.h and sysinfo.cc, therefore this
renaming is innocuous, although unintended.

But if string_util.h is moved above Shlwapi.h then errors occur because
StrCat is not renamed in string_util.h but are renamed in sysinfo.cc:

  ClCompile:
    C:\Program Files (x86)\Microsoft Visual Studio\2017\Community\VC\Tools\MSVC\14.12.25827\bin\HostX86\x64\CL.exe /c /I"C:\Users\winks\prgs\google-benchmark\include" /I"C:\Users\winks\prgs\google-benchmark\src" /I"C:\Users\winks\prgs\google-benchmark\src\..\include" /Zi /nologo /W4 /WX- /diagnostics:classic /Od /Ob0 /D WIN32 /D _WINDOWS /D _CRT_SECURE_NO_WARNINGS /D HAVE_STD_REGEX /D HAVE_STEADY_CLOCK /D "CMAKE_INTDIR=\"Debug\"" /D _MBCS /Gm- /EHsc /RTC1 /MDd /GS /fp:precise /Zc:wchar_t /Zc:forScope /Zc:inline /GR /Fo"benchmark.dir\Debug\\" /Fd"benchmark.dir\Debug\benchmark.pdb" /Gd /TP /errorReport:queue "C:\Users\winks\prgs\google-benchmark\src\sysinfo.cc"
    sysinfo.cc
  C:\Users\winks\prgs\google-benchmark\src\sysinfo.cc(228): error C2660: 'lstrcatA': function does not take 4 arguments [C:\Users\winks\prgs\google-benchmark\build\src\benchmark.vcxproj]
  C:\Users\winks\prgs\google-benchmark\src\sysinfo.cc(229): error C2664: 'LPSTR lstrcatA(LPSTR,LPCSTR)': cannot convert argument 1 from 'std::string' to 'LPSTR' [C:\Users\winks\prgs\google-benchmark\build\src\benchmark.vcxproj]
  ...
  Done Building Project "C:\Users\winks\prgs\google-benchmark\build\test\basic_test.vcxproj" (default targets) -- FAILED.

This PR renames StrCat to StringCat in string_util.h, sysinfo.cc and
console_reporter.cc to fix this issue.