[NFC] Prefix "report(_)?mode" with Aggregation.
This only specifically represents handling of reporting of aggregates.
Not of anything else. Making it more specific makes the name less generic.

This is an issue because i want to add "iteration report mode",
so the naming would be conflicting.

I have split this off, since nobody likes huge commits :)