Apply the `real_time_used >= 5 * min_time` check to manual time as well.
In the commit message of 46afd8e6, I found out the following:
    However when a benchmark uses a manual timer this heuristic isn't helpful
    and likely isn't correct since we don't know what the manual timer actually
    measures.

I think the heuristic is helpful in exactly the same way for manual timer.

In my particular use case with the manual timer, the benchmarks are GPU
benchmarks, and the manual timer reports GPU time. In a certain benchmark,
the real time elapsed 257x more than the manual time. As a result,
ShouldReportIterationResults takes 257x of time to return true.

I think it's reasonable to assume that all users want a benchmark to finish
in a fixed amount of real time for good user experience, regardless of the
real_time / manual_time ratio.