ComputeStats(): actually properly adjust for changed 'iteration' count.
This is a rather huge mess. What is worse, the custom counters
are already divided by iteration count by now.

And i'm not sure it is in general correct to compute
the stats over sum(iterations), and then divide the stat by iteration
count.

I think we should cleanup everything to divide by iteration count
as early as possible.