Update build instructions to better use CMake
Build instructions needlessly referred to make when CMake offers
a command-line interface to abstract away from the specific build
system.

Furthermore, CMake offers command-line "tool mode" which performs basic
filesystem operations. While the syntax is a bit more verbose than
Linux commands it is platform-independent. Now the commands can be
copy-pasted on both Linux and Windows and will just work.

Finally, the Release build type and link-time optimizations are
included in initial commands. A natural flow for a new-comer is to read
and execute the commands and only then learn that one has to go back
and redo them again this time with proper parameters. Now instead the
parameters are only explained later but present already in the initial
commands.