We should keep the information about the run types.
This makes it easier for third-party tools to evaluate if JSON entry
is an aggregate (post-mortem calculation) or iteration (measurement).

E.g. plotting tool may decide to plot the mean values of consecutive
runs. Or measurements only. Or everything. Another example would be
a CI tool which recognizes potential regressions automatically.